Create file for statements utils
